Mergear a develop el sprint 2

+++ src/main/java/com/habitame/api/user/repository/UserRepository.java
@@ -7,18 +7,20 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
     Optional<UserEntity> findByUsername(String username);
+
     Optional<UserEntity> findByEmail(String email);
 
     @Query("""
-        SELECT u FROM UserEntity u
-        WHERE u.username = :value OR u.email = :value
-    """)
+                SELECT u FROM UserEntity u
+                WHERE u.username = :value OR u.email = :value
+            """)
     Optional<UserEntity> findByUsernameOrEmail(@Param("value") String value);
 
     boolean existsByUsername(String username);
+
     boolean existsByEmail(String email);
 }
